feat: Refactor shared chart props
