+++ front-end-components/src/components/charts/types.tsx
@@ -1,4 +1,5 @@
-import React from "react";
+import React, { SVGProps } from "react";
+import { CartesianTickItem } from "recharts/types/util/types";
 
 export type HorizontalBarChartWrapperProps = {
   chartName: string;
@@ -22,4 +23,41 @@ export type ChartDataPoint = {
 export type ChartSortMode = {
   dataPoint: Exclude<keyof ChartDataPoint, "additionalData">;
   direction: "asc" | "desc";
+};
+
+export interface ChartProps<TData extends ChartDataSeries> {
+  chartName: string;
+  data: TData[];
+  grid?: boolean;
+  keyField: keyof TData;
+  legend?: boolean;
+  margin?: number;
+  multiLineAxisLabel?: boolean;
+  seriesConfig?: ChartSeriesConfig<TData>;
+  seriesLabel?: string;
+  seriesLabelField: keyof TData;
+  valueLabel?: string;
+  valueUnit?: ChartSeriesValueUnit;
+}
+
+type ChartSeriesConfig<TData extends ChartDataSeries> = Partial<
+  Record<
+    keyof TData,
+    {
+      className?: string;
+      label?: string;
+      visible: boolean;
+    }
+  >
+>;
+
+type ChartSeriesName = string;
+export type ChartSeriesValue = string | number;
+type ChartSeriesValueUnit = "%";
+export type ChartDataSeries = { [name: ChartSeriesName]: ChartSeriesValue };
+
+export type TickProps = SVGProps<SVGGElement> & {
+  index: number;
+  payload: CartesianTickItem;
+  visibleTicksCount: number;
 };
