+++ front-end-components/src/components/charts/vertical-bar-chart/types.tsx
@@ -1,45 +1,11 @@
-import { SVGProps } from "react";
-import { CartesianTickItem } from "recharts/types/util/types";
+import { ChartDataSeries, ChartProps, ChartSeriesValue } from "src/components";
 
-export interface ChartProps<TData extends ChartDataSeries> {
-  chartName: string;
-  data: TData[];
-  grid?: boolean;
+export interface VerticalBarChartProps<TData extends ChartDataSeries>
+  extends ChartProps<TData> {
   highlightedItemKeys?: ChartSeriesValue[];
-  keyField: keyof TData;
-  legend?: boolean;
-  margin?: number;
-  multiLineAxisLabel?: boolean;
-  ref?: ChartHandler;
-  seriesConfig?: ChartSeriesConfig<TData>;
-  seriesLabel?: string;
-  seriesLabelField: keyof TData;
-  valueLabel?: string;
-  valueUnit?: ChartSeriesValueUnit;
+  ref?: VerticalBarChartHandler;
 }
 
-type ChartSeriesConfig<TData extends ChartDataSeries> = Partial<
-  Record<
-    keyof TData,
-    {
-      className?: string;
-      label?: string;
-      visible: boolean;
-    }
-  >
->;
-
-type ChartSeriesName = string;
-type ChartSeriesValue = string | number;
-type ChartSeriesValueUnit = "%";
-export type ChartDataSeries = { [name: ChartSeriesName]: ChartSeriesValue };
-
-export type ChartHandler = {
+export type VerticalBarChartHandler = {
   download: () => void;
-};
-
-export type TickProps = SVGProps<SVGGElement> & {
-  index: number;
-  payload: CartesianTickItem;
-  visibleTicksCount: number;
 };
